refactor aql effects instances location
