+++ modules/aql/arrow-query-language/build.gradle
@@ -1,9 +1,6 @@
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile project(':arrow-annotations')
-    compile project(':arrow-effects-extensions')
-    compile project(':arrow-effects-reactor-extensions')
-    compile project(':arrow-effects-rx2-extensions')
     compile project(':arrow-mtl')
     kapt project(':arrow-meta')
     kaptTest project(':arrow-meta')
