+++ modules/effects/arrow-effects-extensions/build.gradle
@@ -4,6 +4,7 @@ dependencies {
     compile project(':arrow-annotations')
     compile project(':arrow-extras-extensions')
     compile project(':arrow-mtl')
+    compile project(':arrow-query-language')
     kapt project(':arrow-meta')
     kaptTest project(':arrow-meta')
     compileOnly project(':arrow-meta')
