+++ modules/effects/arrow-effects-reactor-extensions/src/main/kotlin/arrow/effects/reactor/extensions/fluxk.kt
@@ -1,12 +1,17 @@
 package arrow.effects.reactor.extensions
 
 import arrow.Kind
+import arrow.aql.*
 import arrow.core.Either
 import arrow.core.Eval
 import arrow.core.Option
 import arrow.effects.reactor.FluxK
 import arrow.effects.reactor.FluxKOf
 import arrow.effects.reactor.ForFluxK
+import arrow.effects.reactor.extensions.fluxk.applicative.applicative
+import arrow.effects.reactor.extensions.fluxk.foldable.foldable
+import arrow.effects.reactor.extensions.fluxk.functor.functor
+import arrow.effects.reactor.extensions.fluxk.functorFilter.functorFilter
 import arrow.effects.reactor.extensions.fluxk.monad.monad
 import arrow.effects.reactor.extensions.fluxk.monadDefer.monadDefer
 import arrow.effects.reactor.extensions.fluxk.monadError.monadError
@@ -185,4 +190,44 @@ fun FluxK.Companion.monadErrorSwitch(): FluxKMonadError = object : FluxKMonadErr
 interface FluxKFx : Fx<ForFluxK> {
   override fun monadDefer(): MonadDefer<ForFluxK> =
     FluxK.monadDefer()
+}
+
+@extension
+interface FluxKFrom : From<ForFluxK> {
+  override fun applicative(): Applicative<ForFluxK> = FluxK.applicative()
+}
+
+@extension
+interface FluxKSelect : Select<ForFluxK> {
+  override fun functor(): Functor<ForFluxK> = FluxK.functor()
+}
+
+@extension
+interface FluxKWhere : Where<ForFluxK> {
+  override fun functorFilter(): FunctorFilter<ForFluxK> = FluxK.functorFilter()
+}
+
+@extension
+interface FluxKGroupBy : GroupBy<ForFluxK> {
+  override fun foldable(): Foldable<ForFluxK> = FluxK.foldable()
+}
+
+@extension
+interface FluxKCount : Count<ForFluxK> {
+  override fun foldable(): Foldable<ForFluxK> = FluxK.foldable()
+}
+
+@extension
+interface FluxKSum : Sum<ForFluxK> {
+  override fun foldable(): Foldable<ForFluxK> = FluxK.foldable()
+}
+
+@extension
+interface FluxKOrderBy : OrderBy<ForFluxK> {
+  override fun foldable(): Foldable<ForFluxK> = FluxK.foldable()
+}
+
+@extension
+interface FluxKUnion : Union<ForFluxK> {
+  override fun foldable(): Foldable<ForFluxK> = FluxK.foldable()
 }
