+++ modules/effects/arrow-effects-rx2-extensions/build.gradle
@@ -3,6 +3,7 @@ dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile project(':arrow-annotations')
     compile project(':arrow-mtl')
+    compile project(':arrow-query-language')
     kapt project(':arrow-meta')
     kaptTest project(':arrow-meta')
     compileOnly project(':arrow-meta')
